+++ apps/client/src/app/core/database/user.service.ts
@@ -125,21 +125,6 @@ export class UserService extends FirestoreStorage<TeamcraftUser> {
     );
   }
 
-  public set(uid: string, user: TeamcraftUser, ignoreLog = false): Observable<void> {
-    if (!ignoreLog && user.defaultLodestoneId && (user.logProgression.length > 0 || user.gatheringLogProgression.length > 0)) {
-      return this.logTrackingService.set(`${user.$key}:${user.defaultLodestoneId.toString()}`, {
-        crafting: user.logProgression,
-        gathering: user.gatheringLogProgression
-      }).pipe(
-        tap(() => this.recordOperation('write')),
-        switchMap(() => {
-          return super.set(uid, { ...user, gatheringLogProgression: [], logProgression: [] });
-        })
-      );
-    }
-    return super.set(uid, user);
-  }
-
   /**
    * Checks if a given nickname is available.
    * @param {string} nickname