+++ apps/client/src/app/model/user/teamcraft-user.ts
@@ -41,9 +41,6 @@ export class TeamcraftUser extends DataModel {
 
   nickname: string;
 
-  logProgression: number[] = [];
-  gatheringLogProgression: number[] = [];
-
   itemTags: { id: number, tag: string }[] = [];
 
   stats: any = {};