+++ apps/client/src/environments/patch-notes.ts
@@ -1,4 +1,16 @@
 export const patchNotes = `### Bug Fixes
 
-* **desktop:** fixed inventory tracking for some vendor purchases.
-* **lists:** fixed an issue with list copy not working in some cases.`;
+* **db:** fixed a wrong link between some items and recipes having the same id.
+* **db:** fixed broken link from fish db page to fishing spots in gathering details.
+* **db:** fixed gathering masterbook db pages not working properly.
+* **gc-supply:** you can now search without every levels filled.
+* **gearset:** better filtering for ilvl and lvl filters.
+* **log-tracking:** log tracker is now properly split per character.
+* **macro-generator:** fixed an issue with fixed echo not being applied properly.
+* **sidebar:** overlay and social buttons are now out of the scrolling context.
+
+
+### Features
+
+* **gearset:** you can now change job in a gearset (only in same category).
+* **settings:** new autofill filter to only consider NQ items for rows not marked as HQ.`;